Conflicts:
	classes/theme.php

+++ classes/session.php
@@ -58,14 +58,18 @@ class Session
 		self::$lifetime = Plugins::filter( 'session_lifetime', $lifetime );
 		
 
-		session_set_save_handler(
+		$handlers = array(
 			array( 'Session', 'open' ),
 			array( 'Session', 'close' ),
 			array( 'Session', 'read' ),
 			array( 'Session', 'write' ),
 			array( 'Session', 'destroy' ),
-			array( 'Session', 'gc' )
+			array( 'Session', 'gc' ),
 		);
+
+		$handlers = Plugins::filter('session_handlers', $handlers);
+
+		call_user_func_array('session_set_save_handler', $handlers);
 		// session::write gets called after object destruction, so our class isn't available
 		// fix that by registering it as a shutdown function, before objects are destroyed
 		register_shutdown_function( 'session_write_close' );
@@ -128,7 +132,9 @@ class Session
 
 		// Verify expiry
 		if ( HabariDateTime::date_create()->int > $session->expires ) {
-			Session::error( _t( 'Your session expired.' ), 'expired_session' );
+			if ( $session->user_id ) {
+				Session::error( _t( 'Your session expired.' ), 'expired_session' );
+			}
 			$dodelete = true;
 		}
 
@@ -196,7 +202,7 @@ class Session
 			$record = array(
 				'ip' => self::get_subnet( $remote_address ),
 				'expires' => HabariDateTime::date_create()->int + self::$lifetime,
-				'ua' => $user_agent,
+				'ua' => MultiByte::substr( $user_agent, 0, 255 ),
 				'data' => $data,
 			);
 			DB::update(
@@ -243,7 +249,9 @@ class Session
 	 */
 	public static function set_userid( $user_id )
 	{
-		DB::query( 'UPDATE {sessions} SET user_id = ? WHERE token = ?', array( $user_id, session_id() ) );
+		if(!Plugins::filter('session_handlers', false)) {
+			DB::query( 'UPDATE {sessions} SET user_id = ? WHERE token = ?', array( $user_id, session_id() ) );
+		}
 	}
 
 
@@ -253,8 +261,10 @@ class Session
 	 */
 	public static function clear_userid( $user_id )
 	{
-		DB::query( 'DELETE FROM {sessions} WHERE user_id = ? AND token <> ?', array( $user_id, session_id() ) );
-		DB::query( 'UPDATE {sessions} SET user_id = NULL WHERE token = ?', array( session_id() ) );
+		if(!Plugins::filter('session_handlers', false)) {
+			DB::query( 'DELETE FROM {sessions} WHERE user_id = ? AND token <> ?', array( $user_id, session_id() ) );
+			DB::query( 'UPDATE {sessions} SET user_id = NULL WHERE token = ?', array( session_id() ) );
+		}
 	}
 
 	/**
